+++ app/Http/Controllers/Api/SearchController.php
@@ -6,26 +6,34 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\PostResource;
 use App\Http\Resources\TagResource;
 use App\Post;
+use App\Services\ModelSearch;
+use App\Services\Shaarli\Shaarli;
 use App\Tag;
 use Illuminate\Http\Request;
 
 class SearchController extends Controller
 {
-    public function search(Request $request)
+    public function search(Request $request, Shaarli $shaarli)
     {
         $query = $request->get('query');
+        $default_search = $shaarli->getUseDefaultSearch();
 
-        $tags = Tag::search($query)
-                ->take(5)
-                ->get();
+        if (strlen($query) < 3) {
+            abort(422);
+        }
 
-        $posts = Post::search($query)
-            ->query(function ($query) {
+        $tags = (new ModelSearch(Tag::class, ['name']))
+            ->useDefaultSearch($default_search)
+            ->search($query, 5);
+
+
+        $posts = (new ModelSearch(Post::class, ['postable' => ['title', 'content']]))
+            ->useDefaultSearch($default_search)
+            ->withCallback(function ($query) {
                 return $query->with('tags', 'postable')
-                            ->withPrivate(auth('api')->user());
+                    ->withPrivate(auth('api')->user());
             })
-            ->take(10)
-            ->get();
+            ->search($query, 10);
 
         return [
             'posts' => PostResource::collection($posts),