+++ resources/lang/ja/shaarli.php
@@ -9,7 +9,8 @@ return [
             'site_name' => 'サイトの名前',
             'lang' => '言語',
             'private_help' => '非公開（全てのコンテンツが非公開になる。コンテンツへのアクセスにログインが必要。）',
-            'private_download' => 'ダウンロードを非公開にしますか？（ＵＲＬとアルバム）'
+            'private_download' => 'ダウンロードを非公開にしますか？（ＵＲＬとアルバム）',
+            'use_default_search' => 'Use classic SQL search instead of full-text search',
         ],
 
         'appearance' => [